Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    float a, b, c;
+    int a, b, c;
     
-    scanf("%f %f %f", &a, &b, &c);
+    scanf("%d %d %d", &a, &b, &c);
     
     if (a + b > c)
         {printf("Valid\n");}
